server/requirements.txt
openai
google-genai
fastapi
websocket-client
pydantic
uvicorn
dotenv